+++ lib/kitchen/directions/bake_notes/bake_autotitled_notes.rb
@@ -3,20 +3,30 @@
 module Kitchen
   module Directions
     module BakeAutotitledNotes
-      def self.v1(book:, classes:, bake_subtitle: true, cases: false, bake_exercises: false)
+      def self.v1(book:, classes:, options: {
+        bake_subtitle: true,
+        cases: false,
+        bake_exercises: false
+      })
+        options.reverse_merge!(
+          bake_subtitle: true,
+          cases: false,
+          bake_exercises: false
+        )
+
         book.notes.each do |note|
           next unless (note.classes & classes).any?
 
           bake_note(
-            note: note, bake_subtitle: bake_subtitle, cases: cases, bake_exercises: bake_exercises)
+            note: note, options: options)
         end
       end
 
-      def self.bake_note(note:, bake_subtitle:, cases:, bake_exercises:)
+      def self.bake_note(note:, options: {})
         note.wrap_children(class: 'os-note-body')
 
-        if bake_subtitle
-          BakeNoteSubtitle.v1(note: note, cases: cases)
+        if options[:bake_subtitle]
+          BakeNoteSubtitle.v1(note: note, cases: options[:cases])
         else
           note.title&.trash
         end
@@ -29,7 +39,7 @@ module Kitchen
           HTML
         )
 
-        BakeNoteExercise.v2(note: note) if bake_exercises
+        BakeNoteExercise.v2(note: note) if options[:bake_exercises]
       end
     end
   end